+++ src/components/ResetPassword/resetpassword.scss
@@ -20,7 +20,8 @@ $form-color: #f3994a;
 
 	.success {
 		color: green;
-		text-align: left;
+		text-align: center;
+		margin-bottom: 3px;
 	}
 
 	.bg-image {
