+++ src/test/components/card/card.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Card from '../../../components/Card';
+import Card from '../../../components/card';
 
 let Cards;
